Island: Move IAgentBinaryRepository into agent_binary_service
