+++ monkey/monkey_island/cc/services/agent_binary_service/i_agent_binary_repository.py
